--- a/tests/check-dist.spec.ts
+++ b/tests/check-dist.spec.ts
@@ -1,4 +1,4 @@
-import { describe, expect, it } from 'vitest';
+import { beforeAll, describe, expect, it } from 'vitest';
 import { execa } from 'execa';
 import fs from 'node:fs';
 import path from 'node:path';
@@ -8,10 +8,8 @@ import { createTmpDir } from './utils/createTmpDir';
 import { parseTar } from './utils/parseTar';
 import { streamToBuffer } from './utils/streamToBuffer';
 
-async function getPackageJsonOfTarball() {
-  const tarball = await createPackageTarball();
-
-  for await (const entry of parseTar(await fs.promises.readFile(tarball.path))) {
+async function getPackageJsonOfTarball(tarballPath: string) {
+  for await (const entry of parseTar(await fs.promises.readFile(tarballPath))) {
     if (entry.path === 'package/package.json') {
       const json = (await streamToBuffer(entry as unknown as Readable)).toString('utf-8');
 
@@ -38,10 +36,16 @@ const ENTRYPOINTS = [
 ];
 
 describe(`es-toolkit's package tarball`, () => {
+  let tarball: { path: string };
+
+  beforeAll(async () => {
+    tarball = await createPackageTarball();
+  }, 300000);
+
   it(
     'configures all entrypoints correctly',
     async () => {
-      const packageJson = await getPackageJsonOfTarball();
+      const packageJson = await getPackageJsonOfTarball(tarball.path);
       const entrypoints = Object.keys(packageJson.exports);
 
       expect(entrypoints).toEqual([...ENTRYPOINTS, './package.json']);
@@ -52,7 +56,6 @@ describe(`es-toolkit's package tarball`, () => {
   it(
     'exports identical functions in CJS and ESM',
     async () => {
-      const tarball = await createPackageTarball();
       const tmpdir = await createTmpDir();
 
       const packageJson = {